I switched to string lights from strip lights. These lights are WS2811 rather than WS2812B like the strip lights I was using.
I was unable to get the 2811s to work with the Feather HUZZAH I was using.
All kinds of weird refresh rate issues. Random blinking, wrong pixels,
etc. I tried to fix it but ulitmately decided to just use a Trinket M0 I
had lying around. With the Trinket my code works perfectly fine with the
string lights. No idea what was up with the Feather, really.

So anyway this commit is mostly about that. Also changed the flipReverses
function to randomizeReverses instead for greater variety during a
pattern.